added file upload and download using multer, also pdfkit for creating pdf files

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
 
 const { validationResult } = require('express-validator');
@@ -18,19 +20,36 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    console.log(image);
     const errors = validationResult(req);
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: "/admin/add-product", editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not supported!',
+            validationErrors: []
+        });
+    }
+
+    const imageUrl = image.path;
 
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Add Product',
-            path: "/admin/edit-product", editing: false,
+            path: "/admin/add-product", editing: false,
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -45,13 +64,34 @@ exports.postAddProduct = (req, res, next) => {
         description: description,
         userId: req.user
     });
-    product.save()
+    product
+        .save()
         .then(result => {
             // console.log(result);
             console.log('created product');
             res.redirect('/admin/products');
         })
-        .catch(err => console.log(err));
+        .catch(err => {
+            /*
+            return res.status(500).render('admin/edit-product', {
+                pageTitle: 'Add Product',
+                path: "/admin/add-product", editing: false,
+                hasError: true,
+                product: {
+                    title: title,
+                    imageUrl: imageUrl,
+                    price: price,
+                    description: description
+                },
+                errorMessage: 'Database operation failed, please try again.',
+                validationErrors: []
+            });
+            */
+            //res.redirect('/500');
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
 
 
@@ -77,7 +117,9 @@ exports.getEditProduct = (req, res, next) => {
             });
         })
         .catch(err => {
-            console.log(err);
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
         });
 
 };
@@ -86,10 +128,12 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
+
+
 
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
@@ -98,7 +142,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -115,7 +158,12 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
+
             return product
                 .save()
                 .then(result => {
@@ -124,7 +172,9 @@ exports.postEditProduct = (req, res, next) => {
                 });
         })
         .catch(err => {
-            console.log(err);
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
         });
 };
 
@@ -140,20 +190,33 @@ exports.getProducts = (req, res, next) => {
             });
         })
         .catch(err => {
-            console.log(err);
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
         });
 
 };
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
+
     Product
-        .deleteOne({ _id: prodId, userId: req.user._id })
+        .findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product
+                .deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             // console.log('Deleted product');
             res.redirect('/admin/products');
         })
         .catch(err => {
-            console.log(err);
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
         });
 };
